move python world out of agents/test into droidlet/lowlevel/minecraft/pyworld/
