+++ agents/craftassist/tests/test_memory.py
@@ -12,7 +12,8 @@ from droidlet.shared_data_struct.craftassist_shared_utils import MOBS_BY_ID
 from droidlet.interpreter.craftassist import dance
 from droidlet.interpreter.tests.all_test_commands import *
 from agents.craftassist.tests.base_craftassist_test_case import BaseCraftassistTestCase
-from agents.craftassist.tests.utils import Mob, Pos, Look
+from droidlet.lowlevel.minecraft.pyworld.utils import Mob, Pos, Look
+
 
 class ObjectsTest(BaseCraftassistTestCase):
     def setUp(self):
@@ -35,7 +36,9 @@ class ObjectsTest(BaseCraftassistTestCase):
 class TriggersTests(BaseCraftassistTestCase):
     def setUp(self):
         super().setUp()
-        self.cube_right = self.add_object(droidlet.lowlevel.minecraft.shapes.cube(bid=(42, 0)), (9, 63, 4))
+        self.cube_right = self.add_object(
+            droidlet.lowlevel.minecraft.shapes.cube(bid=(42, 0)), (9, 63, 4)
+        )
         self.cube_left = self.add_object(droidlet.lowlevel.minecraft.shapes.cube(), (9, 63, 10))
         self.set_looking_at(list(self.cube_right.blocks.keys())[0])
